# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"L5 OpenApi","description":"L5 Swagger OpenApi description","contact":{"email":"ehs.ghasemi@gmail.com"},"license":{"name":"Apache 2.0","url":"http:\/\/www.apache.org\/licenses\/LICENSE-2.0.html"},"version":"1.0.0"},"paths":{"\/api\/auth\/login":{"post":{"tags":["Auth"],"summary":"Login user using username and password","description":"Returns project data","operationId":"LoginUser","parameters":[{"name":"email","in":"query","description":"User email","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","description":"User password","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"},"400":{"description":"Bad request"},"404":{"description":"Resource Not Found"},"201":{"description":"Successful created","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/api\/auth\/me":{"get":{"tags":["Auth"],"summary":"Get authed user","description":"Returns current authed user","operationId":"Get auth user","responses":{"200":{"description":"successful operation"},"400":{"description":"Bad request"},"404":{"description":"Resource Not Found"},"201":{"description":"Successful created","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/api\/auth\/logout":{"post":{"tags":["Auth"],"summary":"Log the user out (Invalidate the token).","operationId":"LogOut","responses":{"200":{"description":"successful operation"},"400":{"description":"Bad request"},"404":{"description":"Resource Not Found"},"201":{"description":"Successful created","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/api\/auth\/refresh":{"post":{"tags":["Auth"],"summary":"Refresh a token.","operationId":"RefreshToken","responses":{"200":{"description":"successful operation"},"400":{"description":"Bad request"},"404":{"description":"Resource Not Found"},"201":{"description":"Successful created","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/api\/auth\/signup":{"post":{"tags":["Auth"],"summary":"User sign up","operationId":"SignUpUser","parameters":[{"name":"name","in":"query","description":"User name","required":true,"schema":{"type":"string"}},{"name":"family","in":"query","description":"User family","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","description":"User email","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","description":"User password","required":true,"schema":{"type":"string"}},{"name":"password_confirmation","in":"query","description":"User password confirmation","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"},"400":{"description":"Bad request"},"404":{"description":"Resource Not Found"},"201":{"description":"Successful created","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/api\/media":{"post":{"tags":["Media"],"summary":"Store media","operationId":"StorePhoto","requestBody":{"required":true,"content":{"multipart\/form-data":{"schema":{"required":["file"],"properties":{"file":{"description":"file to upload","type":"file","format":"file"}},"type":"object"}}}},"responses":{"200":{"description":"successful operation"},"400":{"description":"Bad request"},"404":{"description":"Resource Not Found"},"201":{"description":"Successful created","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/api\/media\/{media_id}":{"get":{"tags":["Media"],"summary":"Get Preview","description":"Returns Preview of given media id","operationId":"Get Media preview by id","parameters":[{"name":"media_id","in":"path","description":"Media id for file","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"},"400":{"description":"Bad request"},"404":{"description":"Resource Not Found"},"201":{"description":"Successful created","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/api\/product":{"get":{"tags":["Product"],"summary":"Get Product By Category","description":"Get product form category","operationId":"getProduct","parameters":[{"name":"category_id","in":"query","description":"Product category","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"},"400":{"description":"Bad request"},"404":{"description":"Resource Not Found"},"201":{"description":"Successful created","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}}},"components":{"securitySchemes":{"bearer":{"type":"apiKey","description":"A short description for security scheme","bearerFormat":"JWT","name":"Authorization","in":"header"}}}}